requirements.txt
pandas~=1.4.3
streamlit~=1.10.0
scikit-learn~=1.1.1
matplotlib~=3.5.2
seaborn~=0.11.2
plotly~=5.9.0
Pillow~=9.2.0